CXP-1401: moved useCurrencies hook
