+++ components/catalogs/front/src/hooks/useCurrencies.ts
